Convert from ast+types to reflect
This was a failed experiment to be able to use `go/ast` and `go/types`
(via `golang.org/x/tools/go/packages`) instead of compile-then-reflect
for easyjson.  My plan was to write a stupid little package that would
convert from type.Type to reflect types.  It turns out that this is
impossible (at least) because you can't create reflect values for named
types without a value of that type.  In short: the name is a private
field.